ml/requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
pydantic>=2.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
faiss-cpu>=1.7.0
numpy>=1.24.0
python-dotenv>=1.0.0
